refactor: move the movie download callback from the Beatmovies to the App, delete the useEffect from the Search form
