+++ src/pages/Movies/Movies.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useCallback, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 
 import Header from '../../components/Header/Header';
@@ -8,47 +8,19 @@ import MoviesCardList from '../../components/MoviesCardList/MoviesCardList';
 import Preloader from '../../components/Preloader/Preloader';
 import Footer from '../../components/Footer/Footer';
 
-import { MOVIES_ERROR_TEXT } from '../../utils/scripts/constants';
-import { bringMoviesToSingleView, validateMovies } from '../../utils/scripts/utils';
-import getMoviesFromBeatfilm from '../../utils/scripts/MoviesApi';
-
 function Movies({
   savedMovies,
   displayedData,
   setDisplayedData,
+  isProladerShown,
+  setMoviesFromBeatfilm,
   onAddSavedMovie,
   onDeleteSavedMovie,
   onSignOut,
 }) {
-  const [isProladerShown, setIsProladerShown] = useState(false);
-
-  const setMoviesFromBeatfilm = useCallback(async () => {
-    setIsProladerShown(true);
-
-    try {
-      const moviesFromBeatfilm = await getMoviesFromBeatfilm();
-      const convertedMoviesFromBeatfilm = bringMoviesToSingleView(moviesFromBeatfilm);
-      const validatedMoviesFromBeatfilm = validateMovies(convertedMoviesFromBeatfilm);
-
-      localStorage.setItem('moviesFromBeatfilm', JSON.stringify(validatedMoviesFromBeatfilm));
-      setDisplayedData((prevData) => ({
-        ...prevData,
-        allMovies: validatedMoviesFromBeatfilm,
-        queryErrorText: '',
-      }));
-    } catch (error) {
-      setDisplayedData((prevData) => ({
-        ...prevData,
-        allMovies: [],
-        queryErrorText: MOVIES_ERROR_TEXT.FETCH_FAILED,
-      }));
-    }
-
-    setIsProladerShown(false);
-  }, [setDisplayedData]);
-
   useEffect(() => {
     const localMovies = JSON.parse(localStorage.getItem('moviesFromBeatfilm'));
+
     setDisplayedData((prevData) => ({
       ...prevData,
       allMovies: localMovies || [],
@@ -94,6 +66,8 @@ Movies.propTypes = {
     ]),
   ).isRequired,
   setDisplayedData: PropTypes.func.isRequired,
+  isProladerShown: PropTypes.bool.isRequired,
+  setMoviesFromBeatfilm: PropTypes.func.isRequired,
   onAddSavedMovie: PropTypes.func.isRequired,
   onDeleteSavedMovie: PropTypes.func.isRequired,
   onSignOut: PropTypes.func.isRequired,
